services/anomaly-detection-service/requirements.txt
fastapi
uvicorn
kafka-python
redis
numpy
scikit-learn
torch
pydantic
